feat : Add createManualLotto method in LottoGenerator
